k8s: Disable CiliumEndpoint status section by default
The PATCH'ing of the status field of the CiliumEndpoint is often the bottleneck
in scalability. While the status provides value in terms of visibility, it is
not required for any operation.

Disable the CiliumEndpoint status by default to improve the scalability of the
CRD-backed identity allocation mode. Introduce a flag
--enable-endpoint-crd-status to re-enable the status field.

Signed-off-by: Thomas Graf <thomas@cilium.io>